Move data models to core...
...where they probably should have been in the first place.
