+++ sealtk/core/AbstractItemModel.cpp
@@ -2,14 +2,14 @@
  * 3-Clause License. See top-level LICENSE file or
  * https://github.com/Kitware/seal-tk/blob/master/LICENSE for details. */
 
-#include <sealtk/gui/AbstractItemModel.hpp>
+#include <sealtk/core/AbstractItemModel.hpp>
 
 #include <sealtk/core/DataModelTypes.hpp>
 
 namespace sealtk
 {
 
-namespace gui
+namespace core
 {
 
 // ----------------------------------------------------------------------------
@@ -57,6 +57,6 @@ QVariant AbstractItemModel::data(QModelIndex const& index, int role) const
   return {};
 }
 
-} // namespace gui
+} // namespace core
 
 } // namespace sealtk
