+++ sealtk/core/CMakeLists.txt
@@ -11,12 +11,14 @@ set(SEALTK_CORE_VERSION_OUT
 
 sealtk_add_library(sealtk::core
   SOURCES
+    AbstractItemModel.cpp
     AutoLevelsTask.cpp
     DataModelTypes.cpp
     DateUtils.cpp
     FileVideoSourceFactory.cpp
     ImageUtils.cpp
     KwiverFileVideoSourceFactory.cpp
+    KwiverTrackModel.cpp
     KwiverPipelineWorker.cpp
     KwiverVideoSource.cpp
     TimeStamp.cpp
@@ -29,12 +31,14 @@ sealtk_add_library(sealtk::core
     VideoSourceFactory.cpp
 
   HEADERS
+    AbstractItemModel.hpp
     AutoLevelsTask.hpp
     DataModelTypes.hpp
     DateUtils.hpp
     FileVideoSourceFactory.hpp
     ImageUtils.hpp
     KwiverFileVideoSourceFactory.hpp
+    KwiverTrackModel.hpp
     KwiverPipelineWorker.hpp
     KwiverVideoSource.hpp
     TimeMap.hpp
