+++ sealtk/gui/CMakeLists.txt
@@ -6,10 +6,8 @@ include(SEALTKUtils)
 
 sealtk_add_library(sealtk::gui
   SOURCES
-    AbstractItemModel.cpp
     AbstractItemRepresentation.cpp
     Enums.cpp
-    KwiverTrackModel.cpp
     Player.cpp
     PlayerControl.cpp
     Resources.cpp
@@ -18,10 +16,8 @@ sealtk_add_library(sealtk::gui
     resources/Resources.qrc
 
   HEADERS
-    AbstractItemModel.hpp
     AbstractItemRepresentation.hpp
     Enums.hpp
-    KwiverTrackModel.hpp
     Player.hpp
     PlayerControl.hpp
     Resources.hpp
